Add files via upload
This file contains the full code developed for mi Bachelor's Thesis that explores student performance prediction in virtual learning environments using neural networks. The project includes data preprocessing, model training using a feedforward neural network. The application of the NN2Poly interpretability technique to translate the model into an equivalent polynomial form is applied in a second phase, in R (code uplaoded). The aim is to combine predictive accuracy with model transparency. The optimal model and training was implemented in Python and subsequent implementation in R using TensorFlow/Keras and the nn2poly R package.